Update survey strata.geojson
Added an East Scotia Sea (ESS). However, boundaries are slightly tenuous. west boundaries come from the eastern edge of the Scotia Sea (SS) strata. Northern and Southern limits are from the extremes of the SSA, SSB and SSC transects. The area is extended to the east 50 km from the SSA transect, which is what is suggested from Figure 1c (Report of the B0 workshop, La Jolla, USA, 30/05/2000 - 09/06/2000).